// libs/shared/modules/src/data/models/domain/evidenceInstance.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EvidenceInstance {
  /**
   * The unique identifier of this evidence instance
   */
  evidence_instance_id?: string;
  /**
   * The evidence id of this evidence
   */
  evidence_id?: string;
  /**
   * The type of this specific evidence instance
   */
  evidence_type?: EvidenceTypeEnum;
  /**
   * The name of this specific evidence
   */
  evidence_name?: string;
  /**
   * The service this evidence was collected by
   */
  evidence_service_id?: string;

  evidence_originated_by_instance_id?: string;
  /**
   * The display name related with the service that collected this evidence
   */
  evidence_service_display_name?: string;
  /**
   * The time this evidence was collected
   */
  evidence_collection_timestamp?: Date;
  /**
   * Whether this evidence is applicable for the customer
   */
  evidence_is_applicable?: boolean;
  /**
   * Whether this evidence is customized
   */
  evidence_is_custom?: boolean;
  evidence_url?: string;
  evidence_preview?: any;
  evidence_empty_state?: string;
  /**
   * items count in evidence full data
   */
  evidence_items_count?: number;
  /**
   * evidence executed api request
   */
  evidence_api_query?: string[];
  /**
   * evidence time frame, in case of a log type
   */
  evidence_time_frame?: string;
  /**
   * A help phrase describing the purpose of this evidence
   */
  evidence_help?: string;
  evidence_status?: EvidenceStatusEnum;
  evidence_gap?: number[];
  evidence_gap_message?: string;
  evidence_tip?: string;
  evidence_is_beta?: boolean;
  snapshot_id?: string;
}

export type EvidenceTypeEnum = 'DOCUMENT' | 'LOG' | 'LIST' | 'CONFIGURATION' | 'APP' | 'UNKNOWN' | 'LINK' | 'MANUAL' | 'TICKET';
export const EvidenceTypeEnum = {
  DOCUMENT: 'DOCUMENT' as EvidenceTypeEnum,
  LOG: 'LOG' as EvidenceTypeEnum,
  LIST: 'LIST' as EvidenceTypeEnum,
  CONFIGURATION: 'CONFIGURATION' as EvidenceTypeEnum,
  APP: 'APP' as EvidenceTypeEnum,
  UNKNOWN: 'UNKNOWN' as EvidenceTypeEnum,
  LINK: 'LINK' as EvidenceTypeEnum,
  MANUAL: 'MANUAL' as EvidenceTypeEnum,
  URL: 'URL' as EvidenceTypeEnum,
  TICKET: 'TICKET' as EvidenceTypeEnum
};
export type EvidenceStatusEnum = 'NEW' | 'MITIGATED' | 'NOT_MITIGATED';
export const EvidenceStatusEnum = {
  NEW: 'NEW' as EvidenceStatusEnum,
  MITIGATED: 'MITIGATED' as EvidenceStatusEnum,
  NOTMITIGATED: 'NOT_MITIGATED' as EvidenceStatusEnum,
};